requirements.txt
Flask
flask-cors
openai
requests
python-dotenv
